requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlmodel==0.0.14
feedparser==6.0.11
sentence-transformers==2.5.1
python-dotenv==1.0.1
pydantic==2.6.1
apscheduler==3.10.4
numpy==1.26.4
scikit-learn==1.4.0
pytest==8.0.0 